backend/requirements.txt
click==7.1.2
fastapi==0.60.1
h11==0.9.0
pydantic==1.6.1
python-dotenv==0.14.0
starlette==0.13.6
uvicorn==0.11.8
websockets==8.1
